Move Scancode into separate file
